+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Please, check out guidelines: https://keepachangelog.com/en/1.0.0/
 
 - Changed default value of SWIFT_VERSION to 4.2 @ollieatkinson
 - Added fixture tests for ios app with static libraries @ollieatkinson
+- Bundle id linting failing when the bundle id contains variables https://github.com/tuist/tuist/pull/252 by @pepibumur
 
 ## 0.11.0